+++ examples/fw_ike_crypto_profile_add.yml
@@ -1,0 +1,24 @@
+- hosts: my-firewall
+  connection: local
+  gather_facts: False
+
+#  roles:
+#    - role: PaloAltoNetworks.paloaltonetworks
+
+  tasks:
+  - name: Grab the credentials from ansible-vault
+    include_vars: 'firewall-secrets.yml'
+    no_log: 'yes'
+
+  - name: Add IKE crypto config to the firewall
+    panos_ike_crypto_profile:
+      ip_address: '{{ ip_address }}'
+      username: '{{ username }}'
+      password: '{{ password }}'
+      state: 'present'
+      ike_profile_name: 'IKE-Ansible'
+      ike_dhgroup: 'group2'
+      ike_authentication: 'sha1'
+      ike_encryption: ['aes-256-cbc', '3des']
+      ike_lifetime_sec: '28800'
+      commit: 'False'
